requirements.txt
aiogram
asyncio
asyncpg
beautifulsoup4
bs4
lxml
pycoingecko
python-dotenv
requests
